extract the creations of v1 test data
